Reorganized code to increase modularity Also alphabetized imports
