+++ PEAT/utils/communications.py
@@ -1,0 +1,41 @@
+import rsa # pip install rsa
+
+def encrypt(msg):
+    """Encrypts a message using the controller's public key
+    Gets the RSA public key of the controller. Encodes the message in UTF-8 and encrypts it.
+    Returns the encrypted message.
+
+    Args:
+        msg (str): The message to be encrypted
+
+    Returns:
+        bytes: The encrypted message
+    """
+
+    # `ssh-keygen` to generate RSA key pairs
+    # `ssh-keygen -f /path/to/your/public-key -e -m pem > /path/to/your/public-key{.pem}`
+    # to convert the public key into pem format
+    fkey = open('/Users/anmolsaini/.ssh/controller.pub.pem','rb') # modify path
+    public_key = rsa.PublicKey.load_pkcs1(fkey.read())
+    encoded_msg = msg.encode('utf8')
+    encrypted_msg = rsa.encrypt(encoded_msg, public_key)
+    return encrypted_msg
+
+def decrypt(encrypted_msg):
+    """Decrypts a message using PEAT's private key
+    Gets the RSA private key of PEAT. Decrypts the message and decodes it from UTF-8.
+    Returns the decrypted message.
+
+    Args:
+        encrypted_msg (bytes): The message to be decrypted
+
+    Returns:
+        str: The decrypted message
+    """
+
+    # `ssh-keygen -p -m PEM -f /path/to/your/private-key` to convert the private key into pem format
+    fkey = open('/Users/anmolsaini/.ssh/peat', 'rb') # modify path
+    private_key = rsa.PrivateKey.load_pkcs1(fkey.read())
+    decrypted_msg = rsa.decrypt(encrypted_msg, private_key)
+    decoded_msg = decrypted_msg.decode('utf8')
+    return decoded_msg
